Drop unused code and reorganize product utility functions
