--- saleor/product/models/utils.py
@@ -1,15 +1,0 @@
-from django.utils.encoding import smart_text
-
-
-def get_attributes_display_map(obj, attributes):
-    display_map = {}
-    for attribute in attributes:
-        value = obj.attributes.get(smart_text(attribute.pk))
-        if value:
-            choices = {smart_text(a.pk): a for a in attribute.values.all()}
-            choice_obj = choices.get(value)
-            if choice_obj:
-                display_map[attribute.pk] = choice_obj
-            else:
-                display_map[attribute.pk] = value
-    return display_map
